+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.9.0 Unreleased]
+## [1.9.0]
 ###Added
 - Add Google Maps API browser key support (fix #60 #61) Thanks to @DanMan
 - Add possibility to login in the dynamic Google map
 - Add table header customization at location list view in the backend
+- Add dynamic value field in form location list
 
 ###Changed
 - Sorting backend palettes